requirements.txt
flask
pandas
numpy
scikit-learn
joblib
requests
geopy
openai
python-dotenv
